requirements.txt
ibm_cloud_sdk_core==3.16.0
ibm_watson==7.0.0
python-dotenv==1.0.0
